requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.0
PyPDF2==3.0.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
tenacity==8.2.3
azure-functions==1.17.0